created a function for my mountains
after some trial and error of finding out the right code to use to move them over - i asked for some help from copilot on what wasn't working in my code. I has created a variable for the coordinates but this didnt load my code. I was able to identify that I didn't need a variable for such a function and could easily make Mountains its own function calling x,y and having the option of changing these to move the image